os-88: * refactor shoppingCart store slice
