--- src/redux/slices/bucket/bucket.ts
@@ -1,2 +1,0 @@
-export { reducer, name, actions } from './bucket.slice';
-export { selectQuantity } from './selectors';
