+++ src/redux/slices/shopping-cart/selectors.ts
@@ -1,5 +1,5 @@
-import { name as bucketName } from './bucket.slice';
+import { name as shoppingCartName } from './shopping-cart.slice';
 import { type RootState } from '../../store';
 
 export const selectQuantity = (state: RootState): number =>
-  state[bucketName].quantity;
+  state[shoppingCartName].quantity;
