+++ src/redux/slices/shopping-cart/shopping-cart.ts
@@ -1,0 +1,2 @@
+export { reducer, name, actions } from './shopping-cart.slice';
+export { selectQuantity } from './selectors';
